[BUGFIX](project): Renamed / Removed common folder instead of general
Removed common folder and added all files to "General"
